separate terrain uniforms
