+++ www/src/renders/common.ts
@@ -1,10 +1,7 @@
 export class TerrainTextureUniforms {
-    [key: string]: any;
-    constructor() {
-        this.blockSize = 16.0;
-        this.pixelSize = 1.0 / 512.0;
-        this.mipmap = 0;
-    }
+    blockSize = 16.0;
+    pixelSize = 1.0 / 512.0;
+    mipmap = 0;
 
     static default = new TerrainTextureUniforms()
 }
